+++ CHANGES.rst
@@ -13,6 +13,8 @@ The ASDF Standard is at v1.6.0
 - Drop Python 3.8 support [#1556]
 - Drop NumPy 1.20, 1.21 support [#1568]
 - Fix issue opening files that don't support ``fileno`` [#1557]
+- Allow Converters to defer conversion to other Converters
+  by returning ``None`` in ``Converter.select_tag`` [#1561]
 
 2.15.0 (2023-03-28)
 -------------------